Moved vue3-sfc-loader options to pinia store
